backend/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
Flask==2.2.3
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.40.0
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
